+++ src/main/java/com/sparta/newsfeed/common/GlobalExceptionController.java
@@ -3,8 +3,8 @@ package com.sparta.newsfeed.common;
 import com.sparta.newsfeed.common.exception.InvalidUserException;
 import com.sparta.newsfeed.common.exception.NotFoundPostException;
 import com.sparta.newsfeed.common.exception.NotFoundUserException;
-import com.sparta.newsfeed.post.exception.OnlyAuthorAccessException;
-import com.sparta.newsfeed.post.exception.SelfLikeException;
+import com.sparta.newsfeed.post.OnlyAuthorAccessException;
+import com.sparta.newsfeed.like.SelfLikeException;
 import com.sparta.newsfeed.user.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
