+++ src/main/java/com/sparta/newsfeed/like/LikesRepository.java
@@ -8,5 +8,4 @@ import java.util.Optional;
 @Repository
 public interface LikesRepository extends JpaRepository<Likes, Long> {
     Optional<Likes> findByPostIdAndUserId(Long postId, Long userId);
-    int countByPostId(Long postId);
 }
