+++ src/main/java/com/sparta/newsfeed/like/SelfLikeException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.post.exception;
+package com.sparta.newsfeed.like;
 
 public class SelfLikeException extends IllegalArgumentException {
     private static final String MESSAGE = "본인이 작성한 글에 좋아요 할 수 없습니다.";
